+++ Src/AppLib/VpnHood.AppLib.App/Settings/RemoteOptions.cs
@@ -1,0 +1,7 @@
+﻿namespace VpnHood.AppLib.Settings;
+
+public class RemoteOptions
+{
+    public Uri? Url { get; init; }
+    public required RemoteSettings Default{ get; init; }
+}