+++ Src/AppLib/VpnHood.AppLib.App/Settings/RemoteSettings.cs
@@ -1,0 +1,6 @@
+﻿namespace VpnHood.AppLib.Settings;
+
+public class RemoteSettings
+{
+    public Dictionary<string, bool> PreloadAds { get; set; } = [];
+}