+++ Tests/VpnHood.AppLib.Test/Tests/AccessCodeTest.cs
@@ -1,6 +1,4 @@
-﻿using System.Globalization;
-using System.Net.Http.Headers;
-using System.Net.Sockets;
+﻿using System.Net.Sockets;
 using System.Text;
 using VpnHood.AppLib.ClientProfiles;
 using VpnHood.Core.Common.Exceptions;